pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.3",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "python-dateutil>=2.9.0.post0",
    "streamlit>=1.55.0",
    "yfinance>=1.2.0",
]
